backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
python-dotenv
requests